requirements.txt
pyTelegramBotAPI>=4.10.0
python-dotenv>=0.21.0